refactor(database): migrate from SQLite to PostgreSQL
This commit transitions the database from SQLite to PostgreSQL, including updates to the database configuration, migrations, and repository implementations. The changes also introduce a new `Querier` interface to abstract database operations, ensuring compatibility with both `sql.DB` and `sql.Tx`. Additionally, test utilities have been updated to use PostgreSQL via Testcontainers for more accurate testing.

The migration includes:
- Updated `.env.example` with PostgreSQL configuration.
- Modified database initialization and connection logic.
- Adjusted SQL queries to use PostgreSQL syntax.
- Updated test suites to use PostgreSQL for integration testing.
- Introduced a `Querier` interface to abstract database operations.

These changes improve database scalability and maintainability while ensuring compatibility with PostgreSQL-specific features.

+++ internal/monitor/service/target_service_test.go
@@ -64,7 +64,7 @@ func (m *mockNotifierService) Create(notifier *alertModel.Notifier) error {
 	return nil
 }
 
-func (m *mockNotifierService) Get(id int64) (*alertModel.Notifier, error) {
+func (m *mockNotifierService) Get(id int) (*alertModel.Notifier, error) {
 	return nil, nil
 }
 
@@ -72,7 +72,7 @@ func (m *mockNotifierService) Update(id int, config json.RawMessage) (*alertMode
 	return nil, nil
 }
 
-func (m *mockNotifierService) Delete(id int64) error {
+func (m *mockNotifierService) Delete(id int) error {
 	return nil
 }
